django/requirements.txt
django
django-unfold